refactor: change QueryAnswer location
